# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sistema POS Ventas API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: List suppliers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create supplier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "200":
          description: Supplier updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete supplier
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # SALES
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: List sales with optional date filters
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
        - name: customerId
          in: query
          required: false
          schema:
            type: integer
        - name: userId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Process a new sale
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSaleById
      tags: [sales]
      summary: Get sale with details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithDetails"
  /sales/{id}/cancel:
    post:
      operationId: cancelSale
      tags: [sales]
      summary: Cancel sale and restore stock
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # QUOTES
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all system users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # REPORTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required: [id, name, email, role, createdAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
      required: [name, email, password, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: ["string", "null"]
        role:
          type: string
      required: [name, email, role]

    # CATEGORIES
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nitCi:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, createdAt]
    CreateSupplierBody:
      type: object
      properties:
        name:
          type: string
        nitCi:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required: [name]

    # SALES
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
      required: [productId, quantity, unitPrice]
    CreateSaleBody:
      type: object
      properties:
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        paymentMethod:
          type: string
        notes:
          type: ["string", "null"]
      required: [items, paymentMethod]
    SaleDetail:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
      required: [id, productName, quantity, unitPrice, subtotal]
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        subtotal:
          type: number
        iva:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, subtotal, iva, total, paymentMethod, status, createdAt]
    SaleWithDetails:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        subtotal:
          type: number
        iva:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/SaleDetail"
      required: [id, subtotal, iva, total, paymentMethod, status, createdAt, details]
...